fix return type of lame_get_interChRatio()

+++ include/lame.h
@@ -18,7 +18,7 @@
  * the Free Software Foundation, 675 Mass Ave, Cambridge, MA 02139, USA.
  */
 
-/* $Id: lame.h,v 1.110 2002/05/12 17:20:09 markt Exp $ */
+/* $Id: lame.h,v 1.111 2002/06/09 16:01:30 takehiro Exp $ */
 
 #ifndef LAME_LAME_H
 #define LAME_LAME_H
@@ -407,7 +407,7 @@ int CDECL lame_get_useTemporal(const lame_global_flags *);
 
 // use temporal masking effect (default = 1)
 int CDECL lame_set_interChRatio(lame_global_flags *, float);
-int CDECL lame_get_interChRatio(const lame_global_flags *);
+float CDECL lame_get_interChRatio(const lame_global_flags *);
 
 // substep shaping method
 int CDECL lame_set_substep(lame_global_flags *, int);
